(PUP-9247) Update the pidlock process name check for posix platforms
PE's puppet-infrastructure backup and restore system uses
Puppet::Agent::Locker to avoid conflicts with normal runs of the node's
agent. Due to this, the pidlock process name check needs to be expanded
to include process names such as puppet-infrastucture. This loosens the
regex slightly to ensure the lock is still valid when these maintenance
tasks are run.