(PUP-9457) Create connections from an ssl context
Adds a `Puppet::Network::HttpPool.connnection` API allowing connections to be
created from an SSLContext. Using this method is preferred over `http_instance`
and `http_ssl_instance`, because it avoids a number of side effects with the
`Puppet::SSL::Validator::*` classes:

 * The DefaultValidator will lazily lookup the local `Puppet::SSL::Host` object
   to get the client cert and private key.
 * That can cause unexpected cert operations, eg generating a new key, and
   related network connections, eg submitting the CSR, downloading the CRL, etc
 * The validator sometimes gets confused about VERIFY_PEER.
 * The `Puppet::SSL::Host#localhost` method will cause a settings catalog to be
   applied.
 * The default validator's setup_connection method is called for non-SSL
   connections, even though the validator is completely unnecessary.

All of those issues are avoided when using the new API, because all of the
inputs are passed in explicitly via the ssl context.

The `connection` method only accepts an `ssl_context` instead of a `verifier`,
because it's not clear that the latter is required at the moment.