(PUP-9457) Centralize HTTP connection error handling
The `Net::HTTP#start` method can raise an `OpenSSL::SSL::SSLError` if the
server's cert fails verification or if the cert doesn't match the hostname we
tried to connect to.

Previously, the connection pool started the connection, but the `SSLError` was
rescued "far away" in `Puppet::Network::HTTP::Connection`.

This commit creates a `BasePool` class whose start method starts the connection,
and rescues the `SSLError`, if any.

It is possible for an `SSLError` to be raised after the connection is
established. However, `Puppet::Util::SSL.handle_connection_error` previously
re-raised those errors, so moving the rescue closer doesn't change how those
exceptions are handled post-connect.