(PUP-9457) Centralize HTTP connection error handling
The `Net::HTTP#start` method can raise an `OpenSSL::SSL::SSLError` if the any
cert in the server's chain fails verification or if the server's cert doesn't
match the hostname we tried to connect to.

Previously, the connection pool started the connection, but the `SSLError` was
rescued "far away" in `Puppet::Network::HTTP::Connection`.

This commit creates a `BasePool` class whose start method starts the connection,
and rescues the `SSLError`, if any.

It is possible for an `SSLError` exception to be raised after the
`Net::HTTP#start` call, for example, when trying to read from the SSL socket.
Previously, the exception would have been caught by
`Puppet::Util::SSL.handle_connection_error`, but it would have re-raised, since
it wasn't a peer verification error. So moving the rescue closer to the
`Net::HTTP#start` connect call doesn't change how non-connect related exceptions
are handled.