(PUP-9457) Only reuse compatible cached connections
Previously, the `Puppet::Network::HttpPool.connection` method could return a
cached connection whose `ssl_context` was different than what the caller
requested.

This commit ensures we only return cached connections if its verifier is using
the same `ssl_context` instance. Similarly, we don't allow connections created
with a `VerifierAdapter` to be reused when a `Verifier` is requested, and
vice-versa.