(PUP-8586) Improve missing library debug message
This commit re-words the debug message emitted when a library required by a
feature could not be loaded. The new message clarifies that an optional feature
is not being enabled and removes the word "fail" which ensnares folks who are
attempting to debug problems unrelated to the feature system.

Additionally, the debug message is only emitted once per feature and failure
cause.