(maint) Call HttpPool.http_ssl_instance with an integer port
When using `server`, the configurer calls `http_ssl_instance` with an
integer port. When using `server_list`, it calls it with a string. This
works because Puppet::Network::HTTP::Site normalizes the port to an integer.

To avoid future confusion, this commit passes the port as an integer to
match what the docs say, and updates the docs for `http_ssl_instance` to
match that it always returns an integer.