(PUP-9461) Raise meaningful error if SSL bootstrapping is incomplete
If certs, private key, or CRL are not available, then raise an error telling
users which file is missing and to run `puppet agent -t` to bootstrap the SSL
process. For example, if the agent has never run, so there's no CA cert:

    $ puppet filebucket backup file.txt
    Error: Failed to initialize SSL: The CA certificate is missing from '/etc/puppetlabs/puppet/ssl/certs/ca.pem'
    Error: Run `puppet agent -t`

If all of the prerequisite files exist, but we fail to load the SSL context,
such as a mismatched key and cert, then we display why it failed:

    $ puppet catalog find --terminus rest
    Error: Failed to initialize SSL: The certificate for '/CN=chard' does not match its private key
    Error: Run `puppet agent -t`

The user will need to run `puppet agent -t` to complete the bootstrap process.