(PUP-9458) Delete puppet/rest/ssl_context.rb
The Puppet::SSL::SSLContext supersedes the Rest version as it includes
additional fields needed for SSL verification. Delete the latter and update
references to the former.

Note Puppet::SSL::SSLContext defaults `verify_peer` to true, so the only time it
should be specified is when downloading the CA cert.

Also the `Puppet::Rest::Client` class is not currently used, but it's updated to
keep it current.