(PUP-9458) Restrict CA/CRL validation
OpenSSL will raise an error if a PEM file is missing a BEGIN marker[1] or if the
END marker is missing[2]. However, the `load_cacerts_from_pem` and
`load_crls_from_pem` methods would not, which could mask problems later, such as
if we try to fetch the CA certs and the response body is truncated[3].

To protect against that, verify that the CA/CRL PEM contains at least one set of
BEGIN & END markers. Note this doesn't protect against nested markers, eg BEGIN,
BEGIN, END, but I don't think we need that level of validation.

[1] https://github.com/openssl/openssl/blob/OpenSSL_1_1_0j/crypto/pem/pem_lib.c#L686
[2] https://github.com/openssl/openssl/blob/OpenSSL_1_1_0j/crypto/pem/pem_lib.c#L794
[3] https://bugs.ruby-lang.org/issues/14972