(PUP-3905) Disable ruby's http retry
Ruby 2.1.5 introduced automatic retry for idempotent HTTP methods
which ruby defines as: GET HEAD PUT DELETE OPTIONS TRACE. In order to
be idempotent, resending the same request should produce the same
effect as if the request was only sent once.

However, puppetserver processes each report, producing multiple entries in
puppetdb. The latest report doesn't overwrite the earlier report.

Also when puppet submits a report, it will block waiting for the HTTP response
until all report processors are finished. If the connection is dropped, eg load
balancer kills the idle connection, then the agent will retry the request,
causing another instance of the report to be processed.

Ruby 2.5 added an `Net::HTTP#max_retries=` method which defaults to 1, which
means the request will be made once, and retried at most once.

This commit changes the max retry to 0, to disable retry.

In the future, we can add retry logic, but it needs to be done at a higher level
so we can take into account the current REST request and method, such as it's
ok to retry GET file downloads, but not submitting reports.