(PUP-9512) Allow discovering and loading of YAML plans
This adds a second loader for plans which will find plans with a .yaml
extension. Because plans in general, and YAML plans especially, aren't a
meaningful concept in Puppet proper, this loader just calls out to an
injected `yaml_plan_instantiator` to handle actual creation of the plan
once it's been loaded.

This will prefer .pp plans over .yaml when two files with the same name
exist.