(PUP-9511) Log an exec's creates check to debug
Prior to this, an exec's check that the `creates` path exists was hidden
from the user.

That behavior was inconsistent with the `onlyif` and `unless` checks as
those log to debug when they are being checked.

After this commit, when the creates path is being checked for existence,
a debug-level message will be shown to inform the user that the check is
happening.

```
Checking that 'creates' path '/path/to/creates' exists
```