(PUP-9649) Add integration tests for HTTPS client subsystem
Create an HTTPS server for rspec tests. It creates an SSL server listener in a
separate thread, and handles one request at a time. The server expects an HTTP
request, and always replies with HTTP 200 with body "OK".

The server uses a cert whose subject is /CN=127.0.0.1, and it has DNS alt names
for 127.0.0.1 and 127.0.0.2. It's possible to force a cert mismatch error by
connecting to localhost, assuming that is resolvable to 127.0.0.1.

The server is based on the `OpenSSL::SSLTestCase` module in ruby, but simplified
to fit our purposes.

Backport commit bb6468b6531299c943b817971e56fb555bb6c3d7