(PUP-9457) Don't try to stub SSL verification
This commit deletes some SSL related unit tests and replaces them with
integration tests that don't stub the HTTPS connection. The cert
mismatch tests are marked as pending on ruby >= 2.4 due to behavior
changes in ruby[1].

The tests are skipped on JRuby because our certificate error handler
tries to call Open::X509::Certificate#extensions, but the JRuby version
of the class `HTTPClient::JRubySSLSocket::JavaCertificate` doesn't
implement that method.

[1] https://tickets.puppetlabs.com/browse/PUP-8213