(PUP-9337) Add integration tests for File provider on Windows
The logic behind getting and setting the file posix modes can become very
confusing due to how complicated the conversion is.  This commit adds some
integration tests, which uses actual file system calls, to ensure that the
behaviour of the POSIX <-> NTFS ACL conversion is as expected and round-trips
values as much as possible.