(PUP-9459) Implement NeedKey state
Load our private key or generate a new one based on the Puppet[:keylength]
setting. If we load the key successfully, attempt to load the client cert, as it
is in the fast path for each agent run. If the key and cert are mismatched, or
otherwise fails validation, raise an error. If they match, proceed to Done.

If we generate a new key, save it, and proceed to NeedSubmitCSR.