(PUP-9459) Implement NeedSubmitCSR state
Generate and submit a CSR based on the current DNS alt names, CSR
attributes and request extensions. The server will return with 200 if
the PUT request succeeds. It will response with 400 if it already has a
CSR for that name (with the same or different content), if it already
has a signed cert, or the cert is revoked. In all cases, proceed to
NeedCert and attempt to retrieve the client cert.