(PUP-9459) Implement Wait state
If Puppet[:onetime] is set or Puppet[:waitforcert] is less than 1, emit
an error and exit, in the same way the agent does now. Otherwise, wait
`waitforcert` seconds and restart from the beginning of the state
machine. This is necessary because the ssl directory may have been
cleaned while we were sleeping, wiping out the CA certs, our private key, etc