(PUP-8088) Add callbacks at end compilation and environment expiration
This adds callbacks to registered procs when a compilation end and when
an environment expires (environment times out).

In addition, the 4x API now has the ability allowing a function to
register method names to callback for the two "expiration" callback
events. Also added is to get an injected cache object bound to the
lifecycle of the environment.

The ObjectIdCacheAdapter is given a number of helper methods to make use
of a cache easier.

These features taken together allows implementors of functions to handle
cached data / connections with ease.