(PUP-9556) Add puppet ssl bootstrap action
Run the SSL state machine to ensure we have a CA bundle, CRL bundle (if
revocation is enabled), and a private key and signed client cert. This makes it
possible to bootstrap an agent without having to apply a catalog. The bootstrap
command behaves like the agent command, waiting every `waitforcert` seconds if
the cert is not signed. The wait interval can be changed via the setting of the
same name, or disabled by setting it to 0:

    puppet ssl bootstrap --waitforcert 0

In order to see any useful information, set the log level to :info. But don't
lower the log level from :debug, if that's what the user requested.