(PUP-9458) Download CRL explicitly
Previously, if the agent had a private key and cert, but not CA nor CRL bundles,
then the CA bundle was downloaded as a side effect of calling the
Puppet::SSL::Host#certificate method, which occurred during a normal puppet
agent run.

However, the CRL was only downloaded as a side-effect of building an SSL store,
which was called when submitting/retrieving the CSR or when using the
DefaultValidator to verify an SSL connection. If an agent already had a private
key and cert, then the DefaultValidator was solely responsible for downloading
the CRL.

In commit 51311919, puppet's REST terminus was modified to no longer use the
DefaultValidator, instead using Puppet::SSL::Verifier. As a result, puppet no
longer downloaded a CRL if it already had a private key and cert.

This commit modifies Puppet::SSL::Host#certificate to ensure the CRL along with
the CA bundle using the SSL state machine, and deletes unused methods.

Note the tests asserted that ensure_ca_certificate could return false, but in
practice it can't. If the localcacert file exists, then load_certificate_bundle
will return the loaded certs or raise. If the file doesn't exist, then
download_ca_certificate_bundle will return what load_certificate_bundle returns
or raise. As a result the tests were updated to expect an exception and to not
stub Puppet::Rest::Routes behavior.