(PUP-9458) Warn if agent's cert chain is incomplete
If the agent's client cert chain was incomplete, then the agent would fail to
run, breaking the puppetserver test `split_external_cas`.

This commit allows the agent to use an incomplete cert chain, provided the
server has the necessary CA certs locally to complete the chain. It adds
fixtures for an agent-specific intermediate CA, and a client cert for pluto
issued from that CA. If the client cert chain is missing an intermediate CA, a
warning will be generated for the first missing CA, and the run will continue.

Note TLS 1.2 specifies the `Certificate` message MUST contain the complete chain,
except for self-signed root CAs[1]. However, TLS 1.3 relaxes that to say it
SHOULD contain the chain leading up to the trust anchor[2], which could be an
intermediate or self-signed root.

[1] https://tools.ietf.org/html/rfc5246#section-7.4.2
[2] https://tools.ietf.org/html/rfc8446#section-4.4.2