(PUP-9499) Add Puppet::FileSystem.replace_file
Add a method to Puppet::FileSystem for replacing the contents of a file
atomically. It supersedes the Puppet::Util.replace_file method of the same
name, as that version predates Puppet::FileSystem.

If a mode is specified to Puppet::FileSystem.replace_file, then the mode will
always be applied. If a mode is not specified, and the to-be-replaced file
exists, then the mode will be preserved. If the file doesn't exist, the mode
will default to 0640.

In contrast, if a mode is specified to Puppet::Util.replace_file, it only
applies if the file is created, but not when overwriting an existing file.

This commit also ensures Puppet::FileSystem.replace_file behaves consistently
under JRuby when the destination is a directory.