(PUP-9499) Add Windows implementation of FileSystem replace_file
Add a Windows FileSystem replace_file implementation. Like the posix one, if a
mode is specified, it will be applied when creating or overwriting the file.

If a mode is not specified and the file exists, the DACL will be preserved. If
the file doesn't exist, a secure DACL will be applied.

The method only supports translating 3 modes to DACLS. Modes 0600 and 0640
will grant the user, LocalSystem and Administrators group full control. Modes
0644 will do the same, but also grant the builtin Users group read access only.

The commit also raises Errno::EISDIR if the file to be replaced is a directory
and removes the pending label.