(PUP-9463) Set PEM files permissions
When saving a PEM file always set the mode based on the corresponding
puppet setting. The private key is stored as 0640, while the others are
0644.

If we're running as a privileged user, not on Windows, and the service
user/group accounts exist, then set the owner and group when saving the
PEM file. This normally only happens on the puppetserver host, but not
hosts that only have the puppet-agent package. If the accounts do not
exist, then only manage the mode, but not owner or group.