(PUP-9459) Use state machine to ensure client certs
Modify agent and device applications to use the state machine instead of the
Puppet::SSL::Host. Note we no longer have to overwrite `:certname` for each
device.

One change in behavior is `puppet agent --fingerprint` will no longer show the
fingerprint of the CSR since it is generated as needed and never written to
disk.