(PUP-9459) Don't use Puppet::SSL::Host to submit or download
Use the SSLProvider, CertProvider and Rest client to submit the CSR and
download the cert. Now that we always generate and submit a CSR each
time, delete the test verifying whether the previously submitted CSR is
the same as the new one. One outcome of this change is that the agent no
longer retrieves the CSR from the server prior to sending it.