(PUP-4651) Enable services before starting them
Without this, puppet will try to start services before enabling
them. While this was probably fine for sysvinit-style service
managers, some more complex ones (notably systemd) can misbehave when
this is done.

This swaps the order of syncing the enabled property and starting the
service, so that services are enabled before they are started.