(PUP-9459) Disable CRL checking when downloading the CRL
Previously, the connection to download the CRL would fail if revocation checking
was enabled, since we didn't yet have a CRL. The same issue could occur more
subtly if a new CA was downloaded in the CA bundle, but it's CRL wasn't in our
existing CRL (since we hadn't updated it yet).

This commit disables revocation checking for the NeedCRL state, and fixes some
places where the revocation flag wasn't passed through to the SSLContext
correctly.