(PUP-9459) Allow caller to specify if x509 objects are required
The CertProvider#load_* methods are used in cases where we require the object to
exist and cases where we don't. This commit adds a optional `required` keyword
argument. If load an X509 object and it doesn't exist, but is required, then
raise. If it isn't required, return nil as before.

This change means the caller doesn't have to reverse engineer where the cert
provider tried to load the object from, eg `hostcert`, and instead leaves it up
to the cert provider.