(PUP-9579) Push device specific ssl_context
Previously the device application failed when trying to download a
catalog for a device, because the SSL statemachine doesn't take a
`certname` keyword argument.

Also, we have to push the device specific `ssl_context` onto the context
before making network connections on behalf of the device. Otherwise, we
end up using the node's normal ssl cert.