(PUP-9574) Reword allow_duplicate_certs setting
The setting does not allow a cert request to overwrite a certificate. It's more
that the CA will accept a CSR while there is already a pending CSR or a signed
cert with that name. If it's the former, then the new CSR will overwrite the old
one. If it's the latter, then the CSR will be saved, but the user must still
clean (revoke+delete) the old cert, and sign the new CSR.