(PUP-3122) Improve message when cert hasn't been signed yet
Improve message when cert can't be downloaded, so it's clear the agent
is waiting for the cert to be signed on the CA.

Previously the stdout expectation wasn't testing anything due to the
SystemExit. Switch the order of expectations so we rescue the SystemExit
first, and then assert the expected output.