(PUP-8088) Make PAL invoke callbacks for end of env and compile
This completes the implementation of caching/callback support in
functions as demonstrated via tests using PAL.

Tests makes sure that PAL via both script and catalog compilers invokes
the callbacks for end of compilation and eviction of environment.

It was observed that environment.with_override creates a new environment
to be able to set the manifest to use. This was changed so that it is
instead possible to mutate an environment with this information. Without
this change it becomes very difficult to pass the resulting environment
through to the layer where it is known if it expires or not.

The ScriptCompiler is updated to have the same API as the
CatalogCompiler for end of compilation callbacks.