(PUP-9472) Monkey patch OpenSSL::X509::Name#to_utf8
The to_utf8 method was added in ruby 2.5, so monkey patch it to behave
the same on older ruby versions.

As a result of this change the distinguished names in certs and CRLs
will be output in RFC2253 format instead of the broken X509_name_compat
format. Per openssl docs[1]:

    As noted on the ASN1_STRING_print_ex(3) manual page for UTF8
    terminals the ASN1_STRFLGS_ESC_MSB should be unset

For JRuby, don't mask off `ASN1_STRFLGS_ESC_MSB` as it causes the
leading slash to be printed, e.g. /CN=foo.

[1] https://www.openssl.org/docs/man1.1.0/man3/X509_NAME_oneline.html