(PUP-9047) Refactor Puppet.setting usage
Call `Puppet.settings.use` at the right point where all paths are
configured but before any work happens. This will ensure that all
directories are available that puppet requires.

The initial edge case hit was when specifying `--graph` on the CLI or
`graph=true` in the configuration file while running in `--apply` mode,
the `$statedir/graphs` directory was missing.