(PUP-9589) Ignore onetime in state machine
Previously, `puppet agent -t --waitforcert N` ignored `waitforcert`. This was
because `-t` implies `--onetime`, and the state machine would always exit when
`onetime` was set, regardless of `waitforcert`.

This commit simplifies the state machine to only observe the `waitforcert`
setting. The `puppet agent` application already maps `onetime=true` to
`waitforcert=0`, so it doesn't need to be repeated in the state machine.

The `puppet device` application doesn't have the same problem, because
it doesn't have a `--test` option, nor does it automagically map options
to `onetime`.