(maint) define and invoke an `execute_gem_command` method
With this commit ...

The gem provider defines and invokes a common `execute_gem_command` method,
instead of adding `HOME` to `environment` in multiple `execute` method calls.

Adding `HOME` is necessary, as the `gem` command is sensitive to `.gemrc`. And,
if we override any of the options to `execute`, then we have to also re-specify
the defaults for the `failonfail` and `combine` options.

We cannot use the convenience `has_command` method to specify these options ...

`CommandDefiner` in provider.rb creates convenience execution methods:
`commands` and `has_command` that set `failonfail`, `combine`, and optionally,
`environment`. When a child provider defines its own command via `commands`
or `has_command`, the provider-specific path is always returned by `command`.

But when a convenience method is invoked, the last convenience method to be
defined is executed. This makes invoking those convenience methods unsuitable
for inherited providers.

For example, this causes the `puppet_gem` provider to inherit the parent `gem`
provider's convenience `gemcmd` method ... with the wrong path to `gem`.