(PUP-9579) Push device specific ssl_context
Previously the device application failed when trying to download a
catalog for a device, because the SSL statemachine doesn't take a
`certname` keyword argument.

If `--libdir` is specified, then we want to continue skipping pluginsync, but we
still need to ensure the per-device ssl_context is initialized before requesting
a catalog on behalf of the device. If we created a per-device ssl_context, then
push that on the context stack so we don't accidentally fall back to the
non-device one.