(PUP-9497) Allow envioronment_timeout to be Infinity
It is possible that when `environment_timeout = unlimited`, the value
has already been converted to `Infinity` by the time we reach the
deprecation hook. In order to avoid printing out the warning
unexpectedly, we should account for that possibility in the hook itself.
This commit updates the deprecation accordingly.

This commit also similarly updates `filetimeout`. If you set
`filetimeout = 15s`, that string will be recast as the integer `15`, so
we want to allow that value in addition to `15s`.