(PUP-8213) Detect certname mismatch on ruby 2.4/2.5
The Net::HTTP library in ruby 2.4 and up will automatically verify the
server's hostname and cert when VERIFY_PEER is enabled due to changes in
`SSLSocket#connect`. Ruby's openssl bindings will call our verify callback with
`preverify_ok=false` and `store_context.error=0`.

The `Puppet::SSL::Verifier` class already handles this error correctly. It
detects if the current cert causing the error is the leaf cert in the server's
chain, and if the hostname doesn't match the cert.

This commit applies the same logic to the legacy DefaultValidator class so it
behaves the same way. This behavior change will only affect legacy code calling
the old `Puppet::Network::HttpPool.http_instance` or `.ssl_http_instance`
methods, as the newer `.connection` method uses the newer `Verifier`.