(PUP-9055) Add serverless option for compiling a catalog
We removed the `puppet master` application in Puppet 6, but some users relied on
`puppet master --compile` to compile a catalogs without requiring a server.

This adds a `puppet catalog compile` action to compile a catalog for the current
`Puppet[:certname]`. It will collect facts on the local system and include them
in the catalog request, along with the same set of options that the configurer
would send.

Alternative facts may be specified using `--facts <path>`. This is preferable to
`puppet master --compile` which could only load facts for the given hostname
from the `$vardir/yaml/` directory.

Facts may be generated from other nodes using `puppet facts find --render-as yaml`.

The action can load facts if the facts YAML tag is present or not, for example:

    --- !ruby/object:Puppet::Node::Facts