(PUP-9565) Allow specifying --logdest multiple times on the command line
Before PUP-2997 (dbf0ac4f3d) it was possible to specify multiple log
destinations on the command line by repeating the --logdest arg. In order
to restore this ability, we're still setting a flag that lets us know that
we've already set a log destination whenever we handle_logdest_arg, but
handle_logdest_arg no longer bails out early if this flag has been set. We
now use the flag to avoid calling handle_logdest_arg with the "default"
log destination if we've already set the log destination based on command
line arguments.