(PUP-9646) Return parsed yaml plan without instantiating
This returns a parsed yaml plan from the generic plan instantiator without actually instantiating the function. This supports returning a parsed plan to Bolt to be converted into a Puppet plan (see BOLT-1195). It includes loading the plan with a new optional arguments `action` to the relevant loading methods, since we still need to find and load the plan, then uses the action key to determine whether to instantiate or parse the plan. It does not check for whether the plan is a yaml plan or a Puppet plan - should that validation be here or in bolt?