(PUP-9564) Fix package update on resource held
This PR fixes the following bug:

When setting a package to held first it checks if the package is installed on the system, if true then it only sets it's state to held, if false then it installs the latest package and after it sets it's state to held.

Before this it would install the latest package and after that the state would change to held.