(PUP-3237) Accept maximum time to wait for cert
If `waitforcert` is specified, such as when using `puppet agent -t`, then the
agent will wait indefinitely for its cert to be signed, which can be a problem
if there's an infrastructure issue.

This commit adds a `maxwaitforcert` setting of type `ttl`. Acceptable values are
`unlimited` or a duration, such as `5m`, `1h`, etc. The default value is
`unlimited`, so the new behavior must be opted into.