(PUP-8983) Ensure consistent file perms for validate_cmd
Previously, if users had set `validate_cmd` for a file resource, we were
not always guaranteed to have permissions consistent with the final file
resource. Puppet creates and writes to a temporary file, runs the
command specified with `validate_cmd`, then moves the file to its final
destination. This commit ensures that `validate_cmd` is run just before
the file is moved from its temporary to its final location. This gives
us more consistency in the state of the file, including file
permissions, when we run `validate_cmd`.