(PUP-9689) Make Integer.new accept leading 0 for explicit base 10 conv
Before this, the input to Integer.new was checked with a regexp such
that if a value started with a leading zero it must also be a valid
octal number. That was bad because it prevented an explicit conversion
from for example "08" with radix 10.

This is now changes so that leading zeros are accepted by the function,
and that any error is instead triggered when the conversion takes place.

Additional test cases added.