(PUP-9643) Make syntax errors from 3x function loading surface
Before this, if there was a syntax error in a loaded 3.x function
it would be reported as "not containing a call to newfunction" and
the syntax error information would not be shown.

This commits changes this so that syntax errors surface unchanged.

A change was made to the code assertion such that it returns boolean tru
if it was possible to parse, and false otherwise. If false it lets the
real eval of the code string take place and to produce the error.