(PUP-1153) force group removal
The groupdel command from https://github.com/shadow-maint/shadow refuses
to remove groups which are set as the primary GID for a user.  This
effects operating systems such as Debian which uses the shadow codebase.

This is problematic in a Puppet managed world, where it is difficult
to ensure that any dependent users are removed prior to removing the
group.

I added a force option to the groupdel command to force the removal even
when a user has the group set as their GID, so use that flag if it is
available.