(PUP-9693) Add delegation of call_function to Hiera::Scope
Before this interpolation of keys from within a hiera 3 backend where
the lookup of the key specifies `convert_to` would fail because
it tries to call `call_function` on the scope, and scope is a
`Hiera::Scope` wrapping a real scope inside of Hiera 3 code.

This commit adds delegation to the real scope's `call_function` method.