(maint) Exclude *nix service provider tests on Windows and JRuby
Service provider methods such as `instances` for *nix service providers are
never called when running on Windows and JRuby. Similarly, the windows
service provider methods are never called when running on *nix and JRuby.

However, all platforms load the service provider (due to the autoloader), which
triggers the confine logic and ensures the service provider doesn't reference a
platform specific gem, like win32-service, as that would cause compilation to
fail.

So restrict the service provider unit tests to platforms where it makes sense
to test.