(PUP-9653) Fix environmentpath in mk_tmp_env tests
This commit changes tests using the `mk_tmp_env` helper to introspect the
environment path from the puppet `environmentpath` rather than
hard-coding the `/tmp` directory.

This change is necessary to accommodate updates to the helper that ensure
that the `environmentpath` is updated from the temporary environment,
but that ad-hoc changes to the generated temp directory will not be
updated.  This is necessary because Windows cannot link the environment
path to the generated temp directory.