(maint) Remove translation from Ruby version error
The error messaging that occurs when you attempt to load Puppet with an
invalid Ruby version occurs before the i18n machinery is initialized, so
it can't be translated. This commit removes the translation method from
that error.

It also updates the error message to print out the Ruby version that
Puppet found.