(maint) Stub doubles correctly
The `Enumerable#sort` method is a noop when given a single element array. These
tests were relying on that optimization and incorrectly stubbing objects in the
test. Specifically, `ral_spec.rb` wasn't stubbing the `title` and `file_spec.rb`
was returning a string "bar" instead of a `Puppet::Type::File` double.

Update the tests to stub correctly so that we can switch to `sort_by` in a later
commit.