(PUP-9362) Remove redundant tests
The setup data for the test preceding the ones removed structured things
in such a way that all four tests (including the one not removed) had
identical content, and expectations. Rather than run the exact same data
through the exact same process, and expect the exact same results, we only
do this once now, even though it's not (and never was) individually
testing each aspect of the transformation.