(PUP-9717) Retry the state machine when exceptions occur
Prior to 6.4, if an exception occurred, such as failing to connect to the
puppetserver because it's wasn't running yet, then the agent would wait and
retry.

Puppet 6.4 introduced a regression, causing the agent to exit. This commit fixes
the regression. If a StandardError is raised while running the state machine,
then we will wait up to waitforcert seconds and retry the state machine, or exit
if waitforcert is 0.