(PUP-9750) optionally add peserverversion to server_facts
Since PE 3,x, there is no concept of a "PE" Puppet Agent, and therefore
the 'pe_version' fact in 'puppetlabs-stdlib' is nil on platforms > PE 3.x.

Code that varies based upon FOSS vs PE platforms requires a consistent method
to determe whether the platform is FOSS or PE.

Calling a function like 'pe_compiling_server_version' in FOSS will generate an
error in FOSS, and the `is_function_available` function is deprecated,

The concept of a "PE" Puppet Server is valid.

With this commit, Puppet Server sets a 'peserverversion' key in the 'server_facts'
hash of server-side facts for direct inspection, and for use by a future version
of 'puppetlabs-stdlib' that can inspect this key in a best-practice function.